Push more into Window.
Having WidgetList do the caching complicated ownership a little bit, as
it meant it needed to be accessed when either a widget updated its
contents *and* when an X expose event came in. Moving the caching to the
Window (better named as 'Panel' really) seems to make the most sense, as
the updates for the two streams can then be merged there (rather than in
a separate top-level Future as was done previously).

Create our own XcbEvented wrapper rather than trying to use EventedFd
directly. This takes care of ownership/lifetime concerns with the
&RawFd. (I had missed that this was the pattern the docs suggest).

Some general tidying up of Window, in preparation for a subsequent
rename to Panel.