Allow top/bottom bar. Resize based on contents.
Fixes #7, touches #8, but there's some funniness and the bar does not
properly resize if it subsequently changes height. It's probably related
to the oddness around not being able to map the window until it's got a
height. This'll do for now - will keep #8 open and retest once we're no
longer running on QTile, in case it's a bug there.

Also change how text rendering is done - avoid mantaining a
context/layout while working out the geometry. I suspected this may be
causing some of the XCB/Cairo resizing weirdness, but I guess it isn't.
In either case, I think getting rid of TextLayout is a good thing.