Add volume widget.
Rather than shell out to amixer regularly (as QTile's volume widget
does), use alsa-lib to query the volume directly and to register for
events on the relevant mixer.

We don't bother doing anything fancy with the returned events - we just
use them as a wakeup and requery the mixer for its state whenever we
wakeup. (We also make a new Mixer object each time we're woken up, as
the one used for events seems to cache the old state?)

This has the nice side-effect of being much more responsive than the
QTile volume widget I'm currently using *and* much nicer on battery
life. (powertop lists amixer as one of the biggest users of my battery!)

Fixes #2.